direct verification via custom config
