+++ pkg/pb/custom_detectorspb/custom_detectors.pb.validate.go
@@ -344,6 +344,8 @@ func (m *VerifierConfig) validate(all bool) error {
 
 	// no validation rules for HttpMethod
 
+	// no validation rules for DirectVerify
+
 	if len(errors) > 0 {
 		return VerifierConfigMultiError(errors)
 	}
